fix(codex): parent_task_id is bigint to match tasks.id
0078 is still AUTHOR ONLY. The FK type now matches the identity column so apply will not fail.

+++ db/migrations/0078_tasks_help_request.sql
@@ -23,7 +23,7 @@ BEGIN
   END IF;
 
   ALTER TABLE public.tasks
-    ADD COLUMN IF NOT EXISTS parent_task_id integer REFERENCES public.tasks(id) ON DELETE SET NULL,
+    ADD COLUMN IF NOT EXISTS parent_task_id bigint REFERENCES public.tasks(id) ON DELETE SET NULL,
     ADD COLUMN IF NOT EXISTS help_requested_by_email text,
     ADD COLUMN IF NOT EXISTS help_prior_status text;
